Remove dead gate-less AI APIs
