+++ ios-native/MerkenIOS/Services/WebAPIClient.swift
@@ -67,26 +67,6 @@ private struct QuizPrefillResponse: Decodable {
     let error: String?
 }
 
-private struct EmbeddingSyncRequest: Encodable {
-    let wordIds: [String]
-    let limit: Int
-}
-
-private struct EmbeddingSyncResponse: Decodable {
-    let success: Bool?
-    let error: String?
-}
-
-private struct Quiz2SimilarBatchRequest: Encodable {
-    let sourceWordIds: [String]
-    let limit: Int
-}
-
-struct SentenceQuizGeneratedResponse: Sendable {
-    let questions: [SentenceQuizQuestion]
-    let rawResponseData: Data
-}
-
 struct ScanUploadImage: Sendable {
     let data: Data
     let contentType: String
@@ -724,60 +704,6 @@ actor WebAPIClient {
         return words
     }
 
-    func searchSemantic(
-        query: String,
-        bearerToken: String
-    ) async throws -> [SemanticSearchResult] {
-        let url = try makeWebAPIURL(path: "api/search/semantic")
-
-        var request = URLRequest(url: url)
-        request.httpMethod = "POST"
-        request.setValue("application/json", forHTTPHeaderField: "Content-Type")
-        request.setValue("Bearer \(bearerToken)", forHTTPHeaderField: "Authorization")
-        request.timeoutInterval = 30
-
-        let body = ["query": query]
-        request.httpBody = try JSONEncoder().encode(body)
-
-        logger.info("Sending semantic search: query=\(query)")
-
-        let data: Data
-        let response: URLResponse
-        do {
-            (data, response) = try await urlSession.data(for: request)
-        } catch {
-            try rethrowTransportError(error)
-        }
-
-        guard let http = response as? HTTPURLResponse else {
-            throw WebAPIError.serverError("不明な通信エラー")
-        }
-
-        logger.info("Semantic search response: status=\(http.statusCode)")
-
-        switch http.statusCode {
-        case 200...299:
-            break
-        case 401:
-            throw WebAPIError.notAuthenticated
-        case 403:
-            throw WebAPIError.proRequired
-        default:
-            throw WebAPIError.serverError("検索に失敗しました。")
-        }
-
-        let decoded: SemanticSearchResponse
-        do {
-            decoded = try JSONDecoder().decode(SemanticSearchResponse.self, from: data)
-        } catch {
-            logger.error("Semantic search decode failed: \(error.localizedDescription)")
-            throw WebAPIError.decodeFailed(error.localizedDescription)
-        }
-
-        logger.info("Semantic search returned \(decoded.results.count) results")
-        return decoded.results
-    }
-
     func generateQuizPrefill(
         words: [QuizPrefillWordInput],
         bearerToken: String
@@ -820,138 +746,6 @@ actor WebAPIClient {
         }
 
         return decoded.results ?? []
-    }
-
-    func generateSentenceQuizWithRawResponse(
-        words: [SentenceQuizWordInput],
-        bearerToken: String
-    ) async throws -> SentenceQuizGeneratedResponse {
-        logger.info("Sending sentence-quiz request: \(words.count) words")
-
-        let (data, http) = try await sendJSONRequest(
-            path: "api/sentence-quiz/lite",
-            bearerToken: bearerToken,
-            timeout: 30,
-            body: SentenceQuizRequest(words: words, useVectorSearch: false)
-        )
-
-        logger.info("Sentence-quiz response: status=\(http.statusCode)")
-
-        switch http.statusCode {
-        case 200 ... 299:
-            break
-        case 401:
-            throw WebAPIError.notAuthenticated
-        case 403:
-            throw WebAPIError.proRequired
-        case 400:
-            let errorResponse = try? JSONDecoder().decode(SentenceQuizResponse.self, from: data)
-            throw WebAPIError.badRequest(errorResponse?.error ?? "リクエストが不正です。")
-        default:
-            let errorResponse = try? JSONDecoder().decode(SentenceQuizResponse.self, from: data)
-            throw WebAPIError.serverError(errorResponse?.error ?? "サーバーエラーが発生しました。")
-        }
-
-        let decoded: SentenceQuizResponse
-        do {
-            decoded = try JSONDecoder().decode(SentenceQuizResponse.self, from: data)
-        } catch {
-            logger.error("Sentence-quiz decode failed: \(error.localizedDescription)")
-            throw WebAPIError.decodeFailed(error.localizedDescription)
-        }
-
-        guard decoded.success else {
-            throw WebAPIError.serverError(decoded.error ?? "問題の生成に失敗しました。")
-        }
-
-        guard let questions = decoded.questions, !questions.isEmpty else {
-            throw WebAPIError.serverError("問題を生成できませんでした。もう一度お試しください。")
-        }
-
-        logger.info("Generated \(questions.count) sentence-quiz questions")
-        return SentenceQuizGeneratedResponse(questions: questions, rawResponseData: data)
-    }
-
-    func generateSentenceQuiz(
-        words: [SentenceQuizWordInput],
-        bearerToken: String
-    ) async throws -> [SentenceQuizQuestion] {
-        let generated = try await generateSentenceQuizWithRawResponse(
-            words: words,
-            bearerToken: bearerToken
-        )
-        return generated.questions
-    }
-
-    func syncEmbeddings(
-        wordIds: [String],
-        limit: Int,
-        bearerToken: String
-    ) async throws {
-        guard !wordIds.isEmpty else { return }
-        logger.info("Sync embeddings request: \(wordIds.count) words")
-
-        let (data, http) = try await sendJSONRequest(
-            path: "api/embeddings/sync",
-            bearerToken: bearerToken,
-            timeout: 90,
-            body: EmbeddingSyncRequest(wordIds: wordIds, limit: limit)
-        )
-
-        switch http.statusCode {
-        case 200 ... 299:
-            break
-        case 401:
-            throw WebAPIError.notAuthenticated
-        case 403:
-            throw WebAPIError.proRequired
-        case 400:
-            let errorResponse = try? JSONDecoder().decode(EmbeddingSyncResponse.self, from: data)
-            throw WebAPIError.badRequest(errorResponse?.error ?? "Embedding同期リクエストが不正です。")
-        default:
-            let errorResponse = try? JSONDecoder().decode(EmbeddingSyncResponse.self, from: data)
-            throw WebAPIError.serverError(errorResponse?.error ?? "Embedding同期に失敗しました。")
-        }
-
-        if let decoded = try? JSONDecoder().decode(EmbeddingSyncResponse.self, from: data),
-           decoded.success == false {
-            throw WebAPIError.serverError(decoded.error ?? "Embedding同期に失敗しました。")
-        }
-    }
-
-    func warmQuiz2Similar(
-        sourceWordIds: [String],
-        limit: Int,
-        bearerToken: String
-    ) async throws {
-        guard !sourceWordIds.isEmpty else { return }
-        logger.info("Warm quiz2 similar request: \(sourceWordIds.count) words")
-
-        let (data, http) = try await sendJSONRequest(
-            path: "api/quiz2/similar/batch",
-            bearerToken: bearerToken,
-            timeout: 90,
-            body: Quiz2SimilarBatchRequest(sourceWordIds: sourceWordIds, limit: limit)
-        )
-
-        switch http.statusCode {
-        case 200 ... 299:
-            return
-        case 401:
-            throw WebAPIError.notAuthenticated
-        case 403:
-            throw WebAPIError.proRequired
-        case 400:
-            if let message = String(data: data, encoding: .utf8), !message.isEmpty {
-                throw WebAPIError.badRequest(message)
-            }
-            throw WebAPIError.badRequest("類似語キャッシュのリクエストが不正です。")
-        default:
-            if let message = String(data: data, encoding: .utf8), !message.isEmpty {
-                throw WebAPIError.serverError(message)
-            }
-            throw WebAPIError.serverError("類似語キャッシュのウォームアップに失敗しました。")
-        }
     }
 
     func verifyAppStoreTransaction(
