+++ mobile/src/screens/index.ts
@@ -7,4 +7,3 @@ export { LoginScreen } from './LoginScreen';
 export { SignupScreen } from './SignupScreen';
 export { SettingsScreen } from './SettingsScreen';
 export { SubscriptionScreen } from './SubscriptionScreen';
-export { SentenceQuizScreen } from './SentenceQuizScreen';
