--- src/components/sentence-quiz/index.ts
@@ -1,6 +1,0 @@
-export { LoadingScreen } from './LoadingScreen';
-export { QuizProgress } from './QuizProgress';
-export { FillInBlankQuestion } from './FillInBlankQuestion';
-export { MultiFillInBlankQuestion } from './MultiFillInBlankQuestion';
-export { WordOrderQuestion } from './WordOrderQuestion';
-export { QuizResult } from './QuizResult';
