--- src/lib/embeddings/index.ts
@@ -1,5 +1,0 @@
-export {
-  generateWordEmbedding,
-  batchGenerateEmbeddings,
-  EMBEDDING_DIMENSIONS,
-} from './generate';
