+++ src/types/index.ts
@@ -34,24 +34,6 @@ export type {
   Subscription,
   UserState,
   AuthUser,
-  // Sentence Quiz types
-  SentenceQuizType,
-  BlankSlot,
-  FillInBlankQuestion,
-  WordOrderQuestion,
-  SentenceQuizQuestion,
-  AISentenceFillInBlank,
-  AISentenceWordOrder,
-  SentenceQuizWordInput,
-  SentenceQuizResponse,
-  // Enhanced Sentence Quiz types (VectorDB)
-  BlankSource,
-  EnhancedBlankSlot,
-  MultiFillInBlankQuestion,
-  VectorSearchResult,
-  BlankPrediction,
-  AIMultiBlankResponse,
-  EnhancedSentenceQuizQuestion,
   // Collection types
   Collection,
   CollectionProject,
